pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dinotxt"
version = "0.1.0"
description = "Minimal DINOtxt library for loading and inferring DINOtxt vision-language models"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Apache-2.0"}
authors = [
    {name = "vuongnp-eureka", email = "your.email@example.com"}
]
keywords = ["computer-vision", "vision-language", "transformer", "dinov3", "dinotxt"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=1.13.0",
    "torchvision>=0.14.0",
    "numpy>=1.21.0",
    "Pillow>=8.0.0",
    "ftfy>=6.0.0",
    "regex>=2022.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/vuongnp-eureka/dinotxt"
Documentation = "https://github.com/vuongnp-eureka/dinotxt#readme"
Repository = "https://github.com/vuongnp-eureka/dinotxt"
Issues = "https://github.com/vuongnp-eureka/dinotxt/issues"

[tool.setuptools]
packages = ["dinotxt"]

[tool.setuptools.package-data]
dinotxt = ["weights/*.txt.gz"]

